Tidied for gcc-3.4
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@15476 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSCountedSet.m
@@ -206,7 +206,7 @@
 
 - (id) initWithObjects: (id*)objs count: (unsigned)c
 {
-  int i;
+  unsigned int	i;
 
   self = [self initWithCapacity: c];
   if (self == nil)
@@ -273,7 +273,7 @@
 
       while (node != 0)
 	{
-	  if (node->value.uint <= level)
+	  if (node->value.uint <= (unsigned int)level)
 	    {
 	      GSIMapRemoveNodeFromMap(&map, bucket, node);
 	      GSIMapFreeNode(&map, node);
